requirements.txt
anyio==4.10.0
beautifulsoup4==4.12.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
coverage==7.10.6
cycler==0.12.1
ephem==4.2
filelock==3.19.1
fonttools==4.59.2
gdown==5.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==1.26.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2025.2
requests==2.31.0
schedule==1.2.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.0.0
